refactor(backend): move weather context integration to dedicated folder
- Move openMeteoForecastTool.ts to contextIntegrations/weather/
- Move weatherForecastContextStepExecutor.ts to same folder
- Create index.ts as entry point
- Update all import paths across backend
