+++ packages/backend/src/services/contextIntegrations/weather/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @description: Weather context integration entry point.
+ * @footnote-scope: core
+ * @footnote-module: WeatherContextIntegration
+ * @footnote-risk: low - Re-exports only.
+ * @footnote-ethics: low - Re-exports only.
+ */
+export * from './openMeteoForecastTool.js';
+export { createWeatherForecastContextStepExecutor } from './weatherForecastContextStepExecutor.js';
